requirements.txt
asgiref==3.7.2
autoflake==2.2.1
black==23.12.1
click==8.1.7
Django==5.0.1
django-admin-interface==0.28.3
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-decouple==2.1
django-html-field==0.1.3
django-js-asset==2.2.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-decouple==3.8
python-slugify==8.0.1
ruff==0.1.14
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
